Explicitly call `python2` rather than `python`
Whilst on all current stacks `python` is equivalent to `python2`, the
former will not be available on the upcoming Heroku-20, since the
`python2.7` package for Ubuntu 20.04 no longer symlinks it.

This change is a no-op for current stacks, but is landing now to reduce
the size of the Heroku-20 support PR.

In the future we'll want to convert these usages to Python 3, however
there is a non-zero chance of breakage, so it's safer to not try and do
so at the same time as adding support for a new stack.

Refs W-7491249.